Weather service created, methods moved from signage service to clear clutter.
