+++ frontend/src/app/signage/signage.component.html
@@ -6,10 +6,10 @@
             <div class="weather">
                 <img
                 class="weather-image"
-                src="{{assignWeatherIcon(signageService.weatherData())}}"
+                src="{{weatherService.weatherIcon()}}"
                 alt="Weather Icon" />
                 <mat-card-header>
-                    <mat-card-title class="title">{{signageService.weatherData().temperature2m | number: '1.0-0'}}°F</mat-card-title>
+                    <mat-card-title class="title">{{weatherService.weatherData().temperature2m | number: '1.0-0'}}°F</mat-card-title>
                 </mat-card-header>
             </div>
             <div class="announcement">
