+++ frontend/src/assets/weather-icons/foggy.png
