+++ frontend/src/assets/weather-icons/night.png
