+++ frontend/src/assets/weather-icons/overcast.png
