+++ frontend/src/assets/weather-icons/partly-cloudy-windy.png
