+++ frontend/src/assets/weather-icons/partly-cloudy.png
