+++ frontend/src/assets/weather-icons/rainy.png
