+++ frontend/src/assets/weather-icons/snowy.png
