+++ frontend/src/assets/weather-icons/stormy.png
