+++ frontend/src/assets/weather-icons/sunny-windy.png
