+++ frontend/src/assets/weather-icons/sunny.png
